Module ra_syntax::ast::edit

This module contains functions for editing syntax trees. As the trees are immutable, all function here return a fresh copy of the tree, instead of doing an in-place modification.

Structs

IndentLevel

Traits

AstNodeEdit

Functions

remove_attrs_and_docs