Module ra_syntax::ast::make

This module contains free-standing functions for creating AST fragments out of smaller pieces.

Modules

tokens

Functions

add_leading_newlines

add_pub_crate_modifier

add_trailing_newlines

bind_pat

block_expr

block_from_expr

condition

expr_break

expr_continue

expr_empty_block

expr_if

expr_match

expr_path

expr_prefix

expr_return

expr_stmt

expr_todo

expr_unimplemented

expr_unit

fn_def

let_stmt

match_arm

match_arm_list

name

name_ref

param

param_list

path_from_text

path_pat
Returns a BindPat if the path has just one segment, a PathPat otherwise.

path_qualified

path_segment

path_unqualified

placeholder_pat

record_field

record_pat

token

try_expr_from_text

tuple_pat
Creates a tuple of patterns from an interator of patterns.

tuple_struct_pat

unreachable_macro_call

use_item

use_tree

use_tree_list

where_clause

where_pred