Function ra_syntax::lex_single_valid_syntax_kind

pub fn lex_single_valid_syntax_kind(text: &str) -> Option<SyntaxKind>

The same as lex_single_syntax_kind() but returns only SyntaxKind and returns None if any tokenization error occured.

Beware that unescape errors are not checked at tokenization time.