Struct ra_syntax::SyntaxNode

#[repr(transparent)]

pub struct SyntaxNode(_);

Methods

impl SyntaxNode

pub fn kind(&self) -> SyntaxKind

pub fn range(&self) -> TextRange

pub fn text(&self) -> SyntaxText

pub fn parent(&self) -> Option<&SyntaxNode>

pub fn first_child(&self) -> Option<&SyntaxNode>

pub fn first_child_or_token(&self) -> Option<SyntaxElement>

pub fn last_child(&self) -> Option<&SyntaxNode>

pub fn last_child_or_token(&self) -> Option<SyntaxElement>

pub fn next_sibling(&self) -> Option<&SyntaxNode>

pub fn next_sibling_or_token(&self) -> Option<SyntaxElement>

pub fn prev_sibling(&self) -> Option<&SyntaxNode>

pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement>

pub fn children(&self) -> SyntaxNodeChildren

pub fn children_with_tokens(&self) -> SyntaxElementChildren

pub fn first_token(&self) -> Option<SyntaxToken>

pub fn last_token(&self) -> Option<SyntaxToken>

pub fn ancestors(&self) -> impl Iterator<Item = &SyntaxNode>

pub fn descendants(&self) -> impl Iterator<Item = &SyntaxNode>

pub fn descendants_with_tokens(&self) -> impl Iterator<Item = SyntaxElement>

pub fn siblings(
&self,
direction: Direction
) -> impl Iterator<Item = &SyntaxNode>

pub fn siblings_with_tokens(
&self,
direction: Direction
) -> impl Iterator<Item = SyntaxElement>

pub fn preorder(&self) -> impl Iterator<Item = WalkEvent<&SyntaxNode>>

pub fn preorder_with_tokens(
&self
) -> impl Iterator<Item = WalkEvent<SyntaxElement>>

pub fn memory_size_of_subtree(&self) -> usize

pub fn debug_dump(&self) -> String

pub fn insert_children<'a>(
&self,
position: InsertPosition<SyntaxElement>,
to_insert: impl Iterator<Item = SyntaxElement<'a>>
) -> TreeArc<SyntaxNode>

Adds specified children (tokens or nodes) to the current node at the specific position.

This is a type-unsafe low-level editing API, if you need to use it, prefer to create a type-safe abstraction on top of it instead.

pub fn replace_children<'a>(
&self,
to_delete: RangeInclusive<SyntaxElement>,
to_insert: impl Iterator<Item = SyntaxElement<'a>>
) -> TreeArc<SyntaxNode>

Replaces all nodes in to_delete with nodes from to_insert

This is a type-unsafe low-level editing API, if you need to use it, prefer to create a type-safe abstraction on top of it instead.