Struct ra_syntax::TextRange

pub struct TextRange { /* fields omitted */ }

A range in the text, represented as a pair of TextUnits.

Panics

Slicing a &str with TextRange panics if the result is not a valid utf8 string.

Methods

impl TextRange

pub fn checked_sub(self, other: TextUnit) -> Option<TextRange>

impl TextRange

pub fn from_to(from: TextUnit, to: TextUnit) -> TextRange

The left-inclusive range ([from..to)) between to points in the text

pub fn offset_len(offset: TextUnit, len: TextUnit) -> TextRange

The left-inclusive range ([offset..offset + len)) between to points in the text

pub fn start(&self) -> TextUnit

The inclusive start of this range

pub fn end(&self) -> TextUnit

The exclusive end of this range

pub fn len(&self) -> TextUnit

The length of this range

pub fn is_empty(&self) -> bool

Is this range empty of any content?

pub fn is_subrange(&self, other: &TextRange) -> bool

pub fn intersection(&self, other: &TextRange) -> Option<TextRange>

pub fn contains(&self, offset: TextUnit) -> bool

pub fn contains_inclusive(&self, offset: TextUnit) -> bool