Crate ra_syntax

Syntax Tree library used throughout the rust analyzer.

Properties:

easy and fast incremental re-parsing

graceful handling of errors

full-fidelity representation (any text can be precisely represented as a syntax tree)

For more information, see the RFC. Current implementation is inspired by the Swift one.

The most interesting modules here are syntax_node (which defines concrete syntax tree) and ast (which defines abstract syntax tree on top of the CST). The actual parser live in a separate ra_parser crate, thought the lexer lives in this crate.

See api_walkthrough test in this file for a quick API tour!

Re-exports

pub use crate::ast::AstNode;

Modules

algo

ast
Abstract Syntax Tree, layered on top of untyped SyntaxNodes

Structs

AstPtr
Like SyntaxNodePtr, but remembers the type of node

SmolStr
A SmolStr is a string type that has the following properties:

SourceFile

SyntaxError

SyntaxNode

SyntaxNodePtr
A pointer to a syntax node inside a file. It can be used to remember a specific node across reparses of the same file.

SyntaxText

SyntaxToken

SyntaxTreeBuilder

TextRange
A range in the text, represented as a pair of TextUnits.

TextUnit
An offset into text. Offset is represented as u32 storing number of utf8-bytes, but most of the clients should treat it like opaque measure.

Token
A token of Rust source.

TreeArc
An owning smart pointer for CST or AST node.

Enums

Direction

InsertPosition

Location

SyntaxElement

SyntaxErrorKind

SyntaxKind
The kind of syntax node, e.g. IDENT, USE_KW, or STRUCT_DEF.

WalkEvent
WalkEvent describes tree walking process.

Functions

classify_literal

tokenize
Break a string up into its component tokens