Trait ra_syntax::ast::AstNode

pub trait AstNode: TransparentNewType<Repr = SyntaxNode> + ToOwned<Owned = TreeArc<Self>> {
    fn cast(syntax: &SyntaxNode) -> Option<&Self>
    where
        Self: Sized;

    fn syntax(&self) -> &SyntaxNode;
}

The main trait to go from untyped SyntaxNode to a typed ast. The conversion itself has zero runtime cost: ast and syntax nodes have exactly the same representation: a pointer to the tree root and a pointer to the node itself.

Required methods

fn cast(syntax: &SyntaxNode) -> Option<&Self> where
Self: Sized,

fn syntax(&self) -> &SyntaxNode

Loading content...